+++ packages/terra-date-input/tests/wdio/date-input-spec.js
@@ -336,7 +336,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       Terra.it.matchesScreenshot();
     });
 
-    describe('Year Input UP_ARROW increments year by 1 when year value is empty', () => {
+    describe('Year Input UP_ARROW increments year to 1900 when year value is empty', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-date-input/date-input/default-date-input');
         browser.refresh();
@@ -351,7 +351,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       Terra.it.matchesScreenshot();
     });
 
-    describe('Year Input UP_ARROW is cycles to 1 when the year has reached 9999', () => {
+    describe('Year Input UP_ARROW is cycles to 1900 when the year has reached 2100', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-date-input/date-input/default-date-input');
         browser.refresh();
@@ -360,7 +360,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
         Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
 
         browser.click('input[name="terra-date-year-date-input"]');
-        browser.keys(['9999', 'ArrowUp']);
+        browser.keys(['2100', 'ArrowUp']);
       });
 
       Terra.it.matchesScreenshot();
@@ -381,7 +381,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       Terra.it.matchesScreenshot();
     });
 
-    describe('Year Input DOWN_ARROW decrements year to 9999 when year value is empty', () => {
+    describe('Year Input DOWN_ARROW decrements year to 2100 when year value is empty', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-date-input/date-input/default-date-input');
         browser.refresh();
@@ -396,7 +396,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       Terra.it.matchesScreenshot();
     });
 
-    describe('Year Input DOWN_ARROW cycles to 9999 when the year has reached 1', () => {
+    describe('Year Input DOWN_ARROW cycles to 2100 when the year has reached 1900', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-date-input/date-input/default-date-input');
         browser.refresh();
@@ -405,7 +405,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
         Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
 
         browser.click('input[name="terra-date-year-date-input"]');
-        browser.keys(['1', 'ArrowDown']);
+        browser.keys(['1900', 'ArrowDown']);
       });
 
       Terra.it.matchesScreenshot();